Refactored InvokeAsync and Render* methods out of IRenderedComponentBase, changed IRenderedFragment to enable converting method to extension methods
